# Compute jumpdests using the KCFGSemantics.custom_step heuristic