--- a/kevm-pyk/pyproject.toml
+++ b/kevm-pyk/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
 name = "kevm-pyk"
-version = "1.0.567"
+version = "1.0.568"
 description = ""
 authors = [
     "Runtime Verification, Inc. <contact@runtimeverification.com>",