--- a/kevm-pyk/pyproject.toml
+++ b/kevm-pyk/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
 name = "kevm-pyk"
-version = "1.0.568"
+version = "1.0.569"
 description = ""
 authors = [
     "Runtime Verification, Inc. <contact@runtimeverification.com>",