--- a/tests/specs/kontrol/test-allowchangestest-testfailallowchangestostorage-0-spec.k
+++ b/tests/specs/kontrol/test-allowchangestest-testfailallowchangestostorage-0-spec.k
@@ -487,7 +487,7 @@ module TEST-ALLOWCHANGESTEST-TESTFAILALLOWCHANGESTOSTORAGE-0-SPEC
                    b"`\x80`@R4\x80\x15a\x00\x10W`\x00\x80\xfd[P`\xdd\x80a\x00\x1f`\x009`\x00\xf3\xfe`\x80`@R4\x80\x15`\x0fW`\x00\x80\xfd[P`\x046\x10`FW`\x005`\xe0\x1c\x80c\x15\xe6z\xa6\x14`KW\x80c03A;\x14`]W\x80c]3\xa2\x7f\x14`wW\x80cak\x8d\x05\x14`\x7fW[`\x00\x80\xfd[`[`V6`\x04`\x8fV[`\x01UV[\x00[`e`\x00T\x81V[`@Q\x90\x81R` \x01`@Q\x80\x91\x03\x90\xf3[`e`\x01T\x81V[`[`\x8a6`\x04`\x8fV[`\x00UV[`\x00` \x82\x84\x03\x12\x15`\xa0W`\x00\x80\xfd[P5\x91\x90PV\xfe\xa2dipfsX\"\x12 k\xdd\x86\x85\x7f\xfc\x83n#6Y\x8de\xd5'\xc1*\x92U3\x1e\xbcVQ\x12,\xdc\x15\x7fVH\xf5dsolcC\x00\x08\r\x003"
                  </program>
                  <jumpDests>
-                   ( SetItem ( 101 ) ( SetItem ( 106 ) ( SetItem ( 117 ) ( SetItem ( 122 ) ( SetItem ( 124 ) ( SetItem ( 132 ) ( SetItem ( 150 ) ( SetItem ( 158 ) ( SetItem ( 16 ) ( SetItem ( 169 ) ( SetItem ( 174 ) ( SetItem ( 191 ) SetItem ( 46 ) ) ) ) ) ) ) ) ) ) ) ) )
+                   ( SetItem ( 16 ) ( SetItem ( 46 ) ( SetItem ( 117 ) ( SetItem ( 108 ) ( SetItem ( 101 ) ( SetItem ( 106 ) ( SetItem ( 174 ) ( SetItem ( 169 ) ( SetItem ( 160 ) ( SetItem ( 132 ) ( SetItem ( 122 ) ( SetItem ( 124 ) ( SetItem ( 191 ) ( SetItem ( 150 ) SetItem ( 158 ) ) ) ) ) ) ) ) ) ) ) ) ) ) )
                  </jumpDests>
                  <id>
                    ( 491460923342184218035706888008750043977755113263 => 263400868551549723330807389252719309078400616203 )