~-blocked on: https://github.com/runtimeverification/k/pull/4376~
~-blocked on: https://github.com/runtimeverification/evm-semantics/pull/2450~

Implementing a `custom_step` heuristics for KEVMSemantics to compute the jumpdests using kevm-pyk.
Adding the `#loadProgram` as a cut-rule.