@anvacaru 2024-05-23
Ran the test-pyk-prove stage with `--use-booster` active on both this branch and the `v1.0.563`. 

```
| Test                                           | pyjumpdests time | master time | (pyjumpdests/master) time 
|------------------------------------------------|------------------|-------------|---------------------------
| benchmarks/requires01-a0gt0-spec.k             | 37.93            | 42.85       | 0.8851808634772462        
| benchmarks/overflow00-nooverflow-spec.k        | 35.06            | 38.7        | 0.9059431524547804        
| mcd/vow-fess-fail-rough-spec.k                 | 110.27           | 117.9       | 0.9352841391009329        
| bihu/forwardToHotWallet-failure-3-spec.k       | 64.24            | 68.4        | 0.9391812865497075        
| functional/evm-int-simplifications-spec.k      | 76.01            | 80.9        | 0.9395550061804697        
| mcd/vat-move-diff-rough-spec.k                 | 95.97            | 99.87       | 0.9609492340042054        
| functional/int-simplifications-spec.k          | 109.21           | 113.4       | 0.9630511463844796        
| functional/abi-spec.k                          | 84.91            | 87.97       | 0.9652154143457997        
| benchmarks/ecrecover00-siginvalid-spec.k       | 50.68            | 48.91       | 1.0361889184215907        
| benchmarks/ecrecoverloop00-sig0-invalid-spec.k | 54.91            | 52.56       | 1.044710806697108         
| erc20/hkg/totalSupply-spec.k                   | 66.14            | 62.45       | 1.0590872698158527        
| bihu/forwardToHotWallet-success-1-spec.k       | 46.47            | 43.24       | 1.0746993524514339        
| benchmarks/ecrecoverloop00-sigs-valid-spec.k   | 70.3             | 65.28       | 1.0768995098039216        
| benchmarks/ecrecoverloop00-sig1-invalid-spec.k | 103.45           | 94.2        | 1.0981953290870488        
| benchmarks/staticarray00-spec.k                | 39.26            | 34.79       | 1.1284851968956597        
| benchmarks/ecrecover00-sigvalid-spec.k         | 50.91            | 44.83       | 1.1356234664287308        
| TOTAL                                          | 1095.72          | 1096.25     | 0.9995165336374002        

```