@anvacaru 2024-05-30
> Do you think it might make sense to check if all Kontrol tests are still passing with this PR before merging?

I ran the kontrol test suite, and there is no proof getting stuck. However, on the CSE tests it looks like some of the summaries are no longer getting applied and I need to investigate that.